refactor: game package 추가
